t1 pack #5

+++ src/test/java/steps/stateService/StateServiceSteps.java
@@ -182,15 +182,15 @@ public class StateServiceSteps extends Steps {
                 itemData.setFloatingIpAddress(((Map<String, String>) item.getData().get("config")).get("floating_ip_address"));
             else if(item.getType().equals("volume")){
                 itemData.name = ((Map<String, String>) item.getData().get("config")).get("name");
-                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").intValue();
+                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").longValue();
             }
             else if(item.getType().equals("snapshot")){
                 itemData.name = ((Map<String, String>) item.getData().get("config")).get("name");
-                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").intValue();
+                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").longValue();
             }
             else if(item.getType().equals("image")){
                 itemData.name = ((Map<String, String>) item.getData().get("config")).get("name");
-                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").intValue();
+                itemData.size = ((Map<String, Number>) item.getData().get("config")).get("size").longValue();
             }
             else if(item.getType().equals("instance"))
                 itemData.name = ((Map<String, String>) item.getData().get("config")).get("name");
@@ -222,7 +222,7 @@ public class StateServiceSteps extends Steps {
         private String name;
         private String parent;
         private String provider;
-        private Integer size;
+        private Long size;
         private String srcOrderId;
         private String type;
         private String floatingIpAddress;
